feat: evaluate file + kwargs in KD pipeline
